Add tests for Schematron generation.
Note: None of the test have assert/fails yet.

MavenPluginSchematronTest fails because of no namespace prefixes.

XlstSchematronTest succeeds because it has correct namespace prefixes.

PHSchematronTest is an attempt to write tests closer to the guts of where
the code does the work as the MavenPlugin is too high level.
* testPHSchematronDirectlyOutputViaXMLWriter (fails no namespaces)
* testPHSchematronDirectlyOutputViaTransformer (succeeds)
* testPHSchematronDirectlyOutputViaXMLWriterWithNamespaces (an attempt to
configure with correct options to generate correctly)