pure: Add test for #182
This test currently fails because the value of the "var" variable" set
by one thread is overwritten by the other thread.
Theoretically this test would fail or not depending on how the 2 threads
are executed and interweaved. But it seems to always fail, so no further
complication is added for now.